refactor: 더미데이터를 컴포넌트에서 관리하게끔 변경
